backend/requirements.txt
django== 5.1.5
django-environ==0.12.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg==3.2.4
beautifulsoup4==4.12.3
